Separate ParamLookupTable from GridwiseGemmParams.
