+++ mlir/unittests/Dialect/Rock/ParamLookupTableTests.cpp
@@ -1,4 +1,4 @@
-//===- GridwiseGemmParamsTests.cpp - Tests for Tuning Params Lookup -------===//
+//===- ParamLookupTableTests.cpp - Tests for Tuning Params Lookup ---------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
